% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrapping eecop models}
\usage{
bootstrap(object, n_boot = 100, rxi = stats::rexp)
}
\arguments{
\item{object}{a fitted \link{eecop} object.}

\item{n_boot}{number of bootstrap replicates.}

\item{rxi}{a function generating bootstrap multipliers; the function needs to
take the number of samples as its first argument and return a numeric
vector of this length. Default is \code{\link[=rexp]{rexp()}}, which corresponds to the
'Bayesian bootstrap'.}
}
\value{
A list of \link{eecop} objects.
}
\description{
Given a fitted model for the weight function, \code{\link[=bootstrap]{bootstrap()}} generates a
number of bootstrap replicates from this model. A multiplier bootstrap
procedure is used.
}
\examples{
# model with continuous variables
x <- matrix(rnorm(200), 100, 2)
y <- rowSums(x) + rnorm(100)

fit <- eecop(y, x)

bs_fits <- bootstrap(fit, n_boot = 2)
preds <- predict(bs_fits, x[1:3, ])
}
\seealso{
\code{\link[=eecop]{eecop()}}, \code{\link[=predict.eecop_list]{predict.eecop_list()}}
}
