% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eevine.R
\name{deevine}
\alias{deevine}
\alias{dkeeevine}
\title{Working with a \code{eevine} object}
\usage{
deevine(x, object, copula_data = FALSE)
}
\arguments{
\item{x}{\eqn{m x 2} matrix of evaluation points (should be in (0,1) if
\code{copula_data = TRUE}).}

\item{object}{\code{eevine} object.}

\item{copula_data}{if `TRUE``, returns the copula density only.}
}
\value{
A numeric vector for the density.
}
\description{
Working with a \code{eevine} object
}
\examples{
## TODO

}
\seealso{
\code{\link{eevine}}, \code{\link[kdevine]{dkdevine}}
}
