% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eecop-package.R
\docType{package}
\name{eecop-package}
\alias{eecop-package}
\alias{_PACKAGE}
\title{eecop: Copula based estimation equations}
\description{
Solves regression problems by using copula based estimating equations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tnagler/eecop}
  \item Report bugs at \url{https://github.com/tnagler/eecop/issues}
}

}
\author{
\strong{Maintainer}: Thomas Nagler \email{info@vinecopulib.org}

Authors:
\itemize{
  \item Thibault Vatter \email{info@vinecopulib.org}
}

}
\keyword{internal}
