% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eecop.R
\docType{package}
\name{eecop}
\alias{eecop}
\alias{eecop-package}
\title{Copula-Based Solutions to Estimating Equations}
\description{
TODO
}
\examples{
## TODO
}
\author{
Thomas Nagler, Thibault Vatter
}
\keyword{package}
