% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdemcvine.R
\name{kdemcvine}
\alias{kdemcvine}
\title{Multi-class kernel density estimator based on simplified C-vine copulas}
\usage{
kdemcvine(x, y, w, mult_1d = 1, xmin = NULL, xmax = NULL, ...)
}
\arguments{
\item{x}{(\eqn{n x (d-1)}) data matrix.}

\item{y}{(\eqn{n x 1}) data vector for the central node.}

\item{w}{(\eqn{n x 1}) data vector for the classes.}

\item{mult_1d}{numeric; all bandwidhts for marginal kernel density estimation
are multiplied with \code{mult_1d}. Defaults to \code{log(1 + d)}.}

\item{xmin}{numeric vector of length d; see \code{\link[kdevine]{kde1d}}.}

\item{xmax}{numeric vector of length d; see \code{\link[kdevine]{kde1d}}.}

\item{...}{further arguments passed to \code{\link[kdevine]{kde1d}} or
\code{\link[kdevine]{kdevinecop}}.}
}
\value{
An object of class \code{kdemcvine}.
}
\description{
Implements a multi-class C-vine based kernel density estimator.
}
\examples{
TODO

}
\seealso{
\code{\link{dkdecvine}} \code{\link[kdevine]{kdevine}} \code{\link[kdevine]{kdevine}}
}
\author{
Thibault Vatter
}
