% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdemcvine.R
\name{predict.kdemcvine}
\alias{predict.kdemcvine}
\title{Predictions for multi-class kernel based C-vine model}
\usage{
\method{predict}{kdemcvine}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{a \code{kdemcvine} object.}

\item{newdata}{points where the fit shall be evaluated.}

\item{...}{}
}
\value{
A vector of predictions.
}
\description{
Use a fitted multi-class C-vine model to predict the treatment effect.
}
\examples{
TODO

}
\seealso{
\code{\link{kdemcvine}}
}
\author{
Thibault Vatter
}
