% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{solve_eecop}
\alias{solve_eecop}
\title{Generic solver for copula-based estimating equations}
\usage{
solve_eecop(object, x, idfun, theta_start, ...)
}
\arguments{
\item{object}{a fitted \link{eecop} object.}

\item{...}{}
}
\value{

}
\description{
Solves an estimating equation based on a fitted \link{eecop} model and
user-supplied identifying function.
}
