% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.eecop}
\alias{predict.eecop}
\title{Prediction of quantiles or expectiles}
\usage{
\method{predict}{eecop}(object, x, type = "expectile", t = 0.5, trafo = function(y) y, ...)
}
\arguments{
\item{object}{an \code{eecop} object.}

\item{x}{covariate values to predict on; must match the format used for
fitting the \code{eecop()} model.}

\item{type}{either \code{"quantile"}, \code{"expectile"}, \code{"mean"}, or \code{"variance"}.}

\item{t}{a vector of quantile/expectile levels.}

\item{trafo}{a function with signature \verb{function(y)} with \code{y} the
response (vector or matrix). The function should return a
vector of length \code{NROW(object$y)} or a matrix with \code{NROW(object$y)} rows.`}

\item{...}{unused.}
}
\value{
For \code{"quantile"} and \code{"expectile"}: a matrix of predictions, each
column corresponding to one \code{t} (in the order they were supplied to
\code{predict()}. For \code{"mean"}, a matrix containing the predicted mean vectors
in the rows. For  \code{"variance"} an 3-dimensional array containing the
predicted variance/covariance matrices.
}
\description{
Predicts quantiles, expectiles, mean vectors, or covariate matrices from an
\code{eecop()} model by solving a weighted estimating equations as in Nagler and
Vatter (2020).
}
\examples{
# univariate response
x <- matrix(rnorm(200), 100, 2)
y <- rowSums(x) + rnorm(100)

fit <- eecop(y, x)
predict(fit, x[1:3, ], t = c(0.5, 0.9), type = "quantile")
predict(fit, x[1:3, ], t = c(0.5, 0.9), type = "expectile")

# multivariate response
y <- cbind(y1 = y, y2 = y + rnorm(100))
fit <- eecop(y, x)

predict(fit, x[1:3, ], type = "mean")
predict(fit, x[1:3, ], type = "variance")
predict(fit, x[1:3, ], type = "quantile", trafo = function(y) y[, 1] + y[, 2])
}
\references{
Nagler, T. and Vatter, T. (2020). Solving estimating equations
with copulas. arXiv:1801.10576
}
